backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25

# Auth
bcrypt==4.0.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pyotp>=2.9.0
qrcode>=7.4

# Utils
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0
httpx>=0.26.0

# Email
aiosmtplib>=3.0.1

# WebSocket (FastAPI usa starlette, websockets per client lato hub se serve)
websockets>=12.0

# Dev
pytest>=7.4.4
pytest-asyncio>=0.23.3
